+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/SdkLongGaugeBuilder.java
@@ -7,23 +7,23 @@ package io.opentelemetry.sdk.metrics;
 
 import io.opentelemetry.api.metrics.DoubleGaugeBuilder;
 import io.opentelemetry.api.metrics.LongGaugeBuilder;
-import io.opentelemetry.api.metrics.ObservableDoubleMeasurement;
+import io.opentelemetry.api.metrics.ObservableLongMeasurement;
 import io.opentelemetry.sdk.metrics.common.InstrumentType;
 import io.opentelemetry.sdk.metrics.internal.state.MeterProviderSharedState;
 import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
 import java.util.function.Consumer;
 
-final class DoubleGaugeBuilderSdk extends AbstractInstrumentBuilder<DoubleGaugeBuilderSdk>
-    implements DoubleGaugeBuilder {
+final class SdkLongGaugeBuilder extends AbstractInstrumentBuilder<SdkLongGaugeBuilder>
+    implements LongGaugeBuilder {
 
-  DoubleGaugeBuilderSdk(
+  SdkLongGaugeBuilder(
       MeterProviderSharedState meterProviderSharedState,
       MeterSharedState meterSharedState,
       String name) {
     this(meterProviderSharedState, meterSharedState, name, "", "1");
   }
 
-  DoubleGaugeBuilderSdk(
+  SdkLongGaugeBuilder(
       MeterProviderSharedState meterProviderSharedState,
       MeterSharedState sharedState,
       String name,
@@ -33,17 +33,17 @@ final class DoubleGaugeBuilderSdk extends AbstractInstrumentBuilder<DoubleGaugeB
   }
 
   @Override
-  protected DoubleGaugeBuilderSdk getThis() {
+  protected SdkLongGaugeBuilder getThis() {
     return this;
   }
 
   @Override
-  public LongGaugeBuilder ofLongs() {
-    return swapBuilder(LongGaugeBuilderSdk::new);
+  public DoubleGaugeBuilder ofDoubles() {
+    return swapBuilder(SdkDoubleGaugeBuilder::new);
   }
 
   @Override
-  public void buildWithCallback(Consumer<ObservableDoubleMeasurement> callback) {
-    registerDoubleAsynchronousInstrument(InstrumentType.OBSERVABLE_GAUGE, callback);
+  public void buildWithCallback(Consumer<ObservableLongMeasurement> callback) {
+    registerLongAsynchronousInstrument(InstrumentType.OBSERVABLE_GAUGE, callback);
   }
 }
