+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/SdkLongHistogram.java
@@ -6,11 +6,10 @@
 package io.opentelemetry.sdk.metrics;
 
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.metrics.BoundLongCounter;
-import io.opentelemetry.api.metrics.DoubleCounterBuilder;
-import io.opentelemetry.api.metrics.LongCounter;
-import io.opentelemetry.api.metrics.LongCounterBuilder;
-import io.opentelemetry.api.metrics.ObservableLongMeasurement;
+import io.opentelemetry.api.metrics.BoundLongHistogram;
+import io.opentelemetry.api.metrics.DoubleHistogramBuilder;
+import io.opentelemetry.api.metrics.LongHistogram;
+import io.opentelemetry.api.metrics.LongHistogramBuilder;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.common.InstrumentType;
@@ -19,40 +18,36 @@ import io.opentelemetry.sdk.metrics.internal.state.BoundStorageHandle;
 import io.opentelemetry.sdk.metrics.internal.state.MeterProviderSharedState;
 import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
 import io.opentelemetry.sdk.metrics.internal.state.WriteableMetricStorage;
-import java.util.function.Consumer;
 
-final class LongCounterSdk extends AbstractInstrument implements LongCounter {
+final class SdkLongHistogram extends AbstractInstrument implements LongHistogram {
   private final WriteableMetricStorage storage;
 
-  private LongCounterSdk(InstrumentDescriptor descriptor, WriteableMetricStorage storage) {
+  private SdkLongHistogram(InstrumentDescriptor descriptor, WriteableMetricStorage storage) {
     super(descriptor);
     this.storage = storage;
   }
 
   @Override
-  public void add(long increment, Attributes attributes, Context context) {
-    if (increment < 0) {
-      throw new IllegalArgumentException("Counters can only increase");
-    }
-    storage.recordLong(increment, attributes, context);
+  public void record(long value, Attributes attributes, Context context) {
+    storage.recordLong(value, attributes, context);
   }
 
   @Override
-  public void add(long increment, Attributes attributes) {
-    add(increment, attributes, Context.current());
+  public void record(long value, Attributes attributes) {
+    record(value, attributes, Context.current());
   }
 
   @Override
-  public void add(long increment) {
-    add(increment, Attributes.empty());
+  public void record(long value) {
+    record(value, Attributes.empty());
   }
 
   @Override
-  public BoundLongCounter bind(Attributes attributes) {
+  public BoundLongHistogram bind(Attributes attributes) {
     return new BoundInstrument(storage.bind(attributes), attributes);
   }
 
-  static final class BoundInstrument implements BoundLongCounter {
+  static final class BoundInstrument implements BoundLongHistogram {
     private final BoundStorageHandle handle;
     private final Attributes attributes;
 
@@ -62,16 +57,13 @@ final class LongCounterSdk extends AbstractInstrument implements LongCounter {
     }
 
     @Override
-    public void add(long increment, Context context) {
-      if (increment < 0) {
-        throw new IllegalArgumentException("Counters can only increase");
-      }
-      handle.recordLong(increment, attributes, context);
+    public void record(long value, Context context) {
+      handle.recordLong(value, attributes, context);
     }
 
     @Override
-    public void add(long increment) {
-      add(increment, Context.current());
+    public void record(long value) {
+      record(value, Context.current());
     }
 
     @Override
@@ -80,8 +72,8 @@ final class LongCounterSdk extends AbstractInstrument implements LongCounter {
     }
   }
 
-  static final class Builder extends AbstractInstrumentBuilder<Builder>
-      implements LongCounterBuilder {
+  static final class Builder extends AbstractInstrumentBuilder<SdkLongHistogram.Builder>
+      implements LongHistogramBuilder {
 
     Builder(
         MeterProviderSharedState meterProviderSharedState,
@@ -105,19 +97,14 @@ final class LongCounterSdk extends AbstractInstrument implements LongCounter {
     }
 
     @Override
-    public LongCounterSdk build() {
+    public SdkLongHistogram build() {
       return buildSynchronousInstrument(
-          InstrumentType.COUNTER, InstrumentValueType.LONG, LongCounterSdk::new);
+          InstrumentType.HISTOGRAM, InstrumentValueType.LONG, SdkLongHistogram::new);
     }
 
     @Override
-    public DoubleCounterBuilder ofDoubles() {
-      return swapBuilder(DoubleCounterSdk.Builder::new);
-    }
-
-    @Override
-    public void buildWithCallback(Consumer<ObservableLongMeasurement> callback) {
-      registerLongAsynchronousInstrument(InstrumentType.OBSERVABLE_SUM, callback);
+    public DoubleHistogramBuilder ofDoubles() {
+      return swapBuilder(SdkDoubleHistogram.Builder::new);
     }
   }
 }
