+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/SdkLongUpDownCounter.java
@@ -6,10 +6,11 @@
 package io.opentelemetry.sdk.metrics;
 
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.metrics.BoundLongHistogram;
-import io.opentelemetry.api.metrics.DoubleHistogramBuilder;
-import io.opentelemetry.api.metrics.LongHistogram;
-import io.opentelemetry.api.metrics.LongHistogramBuilder;
+import io.opentelemetry.api.metrics.BoundLongUpDownCounter;
+import io.opentelemetry.api.metrics.DoubleUpDownCounterBuilder;
+import io.opentelemetry.api.metrics.LongUpDownCounter;
+import io.opentelemetry.api.metrics.LongUpDownCounterBuilder;
+import io.opentelemetry.api.metrics.ObservableLongMeasurement;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.sdk.metrics.common.InstrumentDescriptor;
 import io.opentelemetry.sdk.metrics.common.InstrumentType;
@@ -18,36 +19,37 @@ import io.opentelemetry.sdk.metrics.internal.state.BoundStorageHandle;
 import io.opentelemetry.sdk.metrics.internal.state.MeterProviderSharedState;
 import io.opentelemetry.sdk.metrics.internal.state.MeterSharedState;
 import io.opentelemetry.sdk.metrics.internal.state.WriteableMetricStorage;
+import java.util.function.Consumer;
 
-final class LongHistogramSdk extends AbstractInstrument implements LongHistogram {
+final class SdkLongUpDownCounter extends AbstractInstrument implements LongUpDownCounter {
   private final WriteableMetricStorage storage;
 
-  private LongHistogramSdk(InstrumentDescriptor descriptor, WriteableMetricStorage storage) {
+  private SdkLongUpDownCounter(InstrumentDescriptor descriptor, WriteableMetricStorage storage) {
     super(descriptor);
     this.storage = storage;
   }
 
   @Override
-  public void record(long value, Attributes attributes, Context context) {
-    storage.recordLong(value, attributes, context);
+  public void add(long increment, Attributes attributes, Context context) {
+    storage.recordLong(increment, attributes, context);
   }
 
   @Override
-  public void record(long value, Attributes attributes) {
-    record(value, attributes, Context.current());
+  public void add(long increment, Attributes attributes) {
+    add(increment, attributes, Context.current());
   }
 
   @Override
-  public void record(long value) {
-    record(value, Attributes.empty());
+  public void add(long increment) {
+    add(increment, Attributes.empty());
   }
 
   @Override
-  public BoundLongHistogram bind(Attributes attributes) {
+  public BoundLongUpDownCounter bind(Attributes attributes) {
     return new BoundInstrument(storage.bind(attributes), attributes);
   }
 
-  static final class BoundInstrument implements BoundLongHistogram {
+  static final class BoundInstrument implements BoundLongUpDownCounter {
     private final BoundStorageHandle handle;
     private final Attributes attributes;
 
@@ -57,13 +59,13 @@ final class LongHistogramSdk extends AbstractInstrument implements LongHistogram
     }
 
     @Override
-    public void record(long value, Context context) {
-      handle.recordLong(value, attributes, context);
+    public void add(long increment, Context context) {
+      handle.recordLong(increment, attributes, context);
     }
 
     @Override
-    public void record(long value) {
-      record(value, Context.current());
+    public void add(long increment) {
+      add(increment, Context.current());
     }
 
     @Override
@@ -72,8 +74,8 @@ final class LongHistogramSdk extends AbstractInstrument implements LongHistogram
     }
   }
 
-  static final class Builder extends AbstractInstrumentBuilder<LongHistogramSdk.Builder>
-      implements LongHistogramBuilder {
+  static final class Builder extends AbstractInstrumentBuilder<SdkLongUpDownCounter.Builder>
+      implements LongUpDownCounterBuilder {
 
     Builder(
         MeterProviderSharedState meterProviderSharedState,
@@ -97,14 +99,19 @@ final class LongHistogramSdk extends AbstractInstrument implements LongHistogram
     }
 
     @Override
-    public LongHistogramSdk build() {
+    public LongUpDownCounter build() {
       return buildSynchronousInstrument(
-          InstrumentType.HISTOGRAM, InstrumentValueType.LONG, LongHistogramSdk::new);
+          InstrumentType.UP_DOWN_COUNTER, InstrumentValueType.LONG, SdkLongUpDownCounter::new);
     }
 
     @Override
-    public DoubleHistogramBuilder ofDoubles() {
-      return swapBuilder(DoubleHistogramSdk.Builder::new);
+    public DoubleUpDownCounterBuilder ofDoubles() {
+      return swapBuilder(SdkDoubleUpDownCounter.Builder::new);
+    }
+
+    @Override
+    public void buildWithCallback(Consumer<ObservableLongMeasurement> callback) {
+      registerLongAsynchronousInstrument(InstrumentType.OBSERVABLE_UP_DOWN_SUM, callback);
     }
   }
 }
