+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/SdkDoubleCounterTest.java
@@ -10,8 +10,8 @@ import static io.opentelemetry.sdk.testing.assertj.metrics.MetricAssertions.asse
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.metrics.BoundDoubleUpDownCounter;
-import io.opentelemetry.api.metrics.DoubleUpDownCounter;
+import io.opentelemetry.api.metrics.BoundDoubleCounter;
+import io.opentelemetry.api.metrics.DoubleCounter;
 import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.metrics.StressTestRunner.OperationUpdater;
@@ -20,13 +20,13 @@ import io.opentelemetry.sdk.testing.time.TestClock;
 import java.time.Duration;
 import org.junit.jupiter.api.Test;
 
-/** Unit tests for {@link DoubleUpDownCounterSdk}. */
-class DoubleUpDownCounterSdkTest {
+/** Unit tests for {@link SdkDoubleCounter}. */
+class SdkDoubleCounterTest {
   private static final long SECOND_NANOS = 1_000_000_000;
   private static final Resource RESOURCE =
       Resource.create(Attributes.of(stringKey("resource_key"), "resource_value"));
   private static final InstrumentationLibraryInfo INSTRUMENTATION_LIBRARY_INFO =
-      InstrumentationLibraryInfo.create(DoubleUpDownCounterSdkTest.class.getName(), null);
+      InstrumentationLibraryInfo.create(SdkDoubleCounterTest.class.getName(), null);
   private final TestClock testClock = TestClock.create();
   private final SdkMeterProvider sdkMeterProvider =
       SdkMeterProvider.builder().setClock(testClock).setResource(RESOURCE).build();
@@ -35,30 +35,22 @@ class DoubleUpDownCounterSdkTest {
   @Test
   void add_PreventNullAttributes() {
     assertThatThrownBy(
-            () ->
-                sdkMeter
-                    .upDownCounterBuilder("testUpDownCounter")
-                    .ofDoubles()
-                    .build()
-                    .add(1.0, null))
+            () -> sdkMeter.counterBuilder("testCounter").ofDoubles().build().add(1.0, null))
         .isInstanceOf(NullPointerException.class)
         .hasMessage("attributes");
   }
 
   @Test
   void bound_PreventNullAttributes() {
-    assertThatThrownBy(
-            () -> sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build().bind(null))
+    assertThatThrownBy(() -> sdkMeter.counterBuilder("testCounter").ofDoubles().build().bind(null))
         .isInstanceOf(NullPointerException.class)
         .hasMessage("attributes");
   }
 
   @Test
   void collectMetrics_NoRecords() {
-    DoubleUpDownCounter doubleUpDownCounter =
-        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
-    BoundDoubleUpDownCounter bound =
-        doubleUpDownCounter.bind(Attributes.builder().put("foo", "bar").build());
+    DoubleCounter doubleCounter = sdkMeter.counterBuilder("testCounter").ofDoubles().build();
+    BoundDoubleCounter bound = doubleCounter.bind(Attributes.builder().put("foo", "bar").build());
     try {
       assertThat(sdkMeterProvider.collectAllMetrics()).isEmpty();
     } finally {
@@ -69,27 +61,27 @@ class DoubleUpDownCounterSdkTest {
   @Test
   @SuppressWarnings("unchecked")
   void collectMetrics_WithEmptyAttributes() {
-    DoubleUpDownCounter doubleUpDownCounter =
+    DoubleCounter doubleCounter =
         sdkMeter
-            .upDownCounterBuilder("testUpDownCounter")
+            .counterBuilder("testCounter")
             .ofDoubles()
             .setDescription("description")
             .setUnit("ms")
             .build();
     testClock.advance(Duration.ofNanos(SECOND_NANOS));
-    doubleUpDownCounter.add(12d, Attributes.empty());
-    doubleUpDownCounter.add(12d);
+    doubleCounter.add(12d, Attributes.empty());
+    doubleCounter.add(12d);
     assertThat(sdkMeterProvider.collectAllMetrics())
         .satisfiesExactly(
             metric ->
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testUpDownCounter")
+                    .hasName("testCounter")
                     .hasDescription("description")
                     .hasUnit("ms")
                     .hasDoubleSum()
-                    .isNotMonotonic()
+                    .isMonotonic()
                     .isCumulative()
                     .points()
                     .satisfiesExactly(
@@ -105,28 +97,28 @@ class DoubleUpDownCounterSdkTest {
   @SuppressWarnings("unchecked")
   void collectMetrics_WithMultipleCollects() {
     long startTime = testClock.now();
-    DoubleUpDownCounter doubleUpDownCounter =
-        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
-    BoundDoubleUpDownCounter bound =
-        doubleUpDownCounter.bind(Attributes.builder().put("K", "V").build());
+    DoubleCounter doubleCounter = sdkMeter.counterBuilder("testCounter").ofDoubles().build();
+    BoundDoubleCounter bound = doubleCounter.bind(Attributes.builder().put("K", "V").build());
     try {
       // Do some records using bounds and direct calls and bindings.
-      doubleUpDownCounter.add(12.1d, Attributes.empty());
+      doubleCounter.add(12.1d, Attributes.empty());
       bound.add(123.3d);
-      doubleUpDownCounter.add(21.4d, Attributes.empty());
+      doubleCounter.add(21.4d, Attributes.empty());
       // Advancing time here should not matter.
       testClock.advance(Duration.ofNanos(SECOND_NANOS));
       bound.add(321.5d);
-      doubleUpDownCounter.add(111.1d, Attributes.builder().put("K", "V").build());
+      doubleCounter.add(111.1d, Attributes.builder().put("K", "V").build());
       assertThat(sdkMeterProvider.collectAllMetrics())
           .satisfiesExactly(
               metric ->
                   assertThat(metric)
                       .hasResource(RESOURCE)
                       .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                      .hasName("testUpDownCounter")
+                      .hasName("testCounter")
+                      .hasDescription("")
+                      .hasUnit("1")
                       .hasDoubleSum()
-                      .isNotMonotonic()
+                      .isMonotonic()
                       .isCumulative()
                       .points()
                       .allSatisfy(
@@ -140,21 +132,19 @@ class DoubleUpDownCounterSdkTest {
                           point ->
                               assertThat(point)
                                   .hasValue(555.9)
-                                  .hasAttributes(Attributes.of(stringKey("K"), "V"))));
+                                  .attributes()
+                                  .hasSize(1)
+                                  .containsEntry("K", "V")));
 
       // Repeat to prove we keep previous values.
       testClock.advance(Duration.ofNanos(SECOND_NANOS));
       bound.add(222d);
-      doubleUpDownCounter.add(11d, Attributes.empty());
+      doubleCounter.add(11d, Attributes.empty());
       assertThat(sdkMeterProvider.collectAllMetrics())
           .satisfiesExactly(
               metric ->
                   assertThat(metric)
-                      .hasResource(RESOURCE)
-                      .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                      .hasName("testUpDownCounter")
                       .hasDoubleSum()
-                      .isNotMonotonic()
                       .isCumulative()
                       .points()
                       .allSatisfy(
@@ -175,26 +165,41 @@ class DoubleUpDownCounterSdkTest {
   }
 
   @Test
+  void doubleCounterAdd_Monotonicity() {
+    DoubleCounter doubleCounter = sdkMeter.counterBuilder("testCounter").ofDoubles().build();
+
+    assertThatThrownBy(() -> doubleCounter.add(-45.77d, Attributes.empty()))
+        .isInstanceOf(IllegalArgumentException.class);
+  }
+
+  @Test
+  void boundDoubleCounterAdd_Monotonicity() {
+    DoubleCounter doubleCounter = sdkMeter.counterBuilder("testCounter").ofDoubles().build();
+
+    assertThatThrownBy(() -> doubleCounter.bind(Attributes.empty()).add(-9.3))
+        .isInstanceOf(IllegalArgumentException.class);
+  }
+
+  @Test
   @SuppressWarnings("unchecked")
   void stressTest() {
-    final DoubleUpDownCounter doubleUpDownCounter =
-        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
+    final DoubleCounter doubleCounter = sdkMeter.counterBuilder("testCounter").ofDoubles().build();
 
     StressTestRunner.Builder stressTestBuilder =
         StressTestRunner.builder()
-            .setInstrument((DoubleUpDownCounterSdk) doubleUpDownCounter)
+            .setInstrument((SdkDoubleCounter) doubleCounter)
             .setCollectionIntervalMs(100);
 
     for (int i = 0; i < 4; i++) {
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
-              1_000, 2, new OperationUpdaterDirectCall(doubleUpDownCounter, "K", "V")));
+              1_000, 2, new OperationUpdaterDirectCall(doubleCounter, "K", "V")));
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
               1_000,
               2,
               new OperationUpdaterWithBinding(
-                  doubleUpDownCounter.bind(Attributes.builder().put("K", "V").build()))));
+                  doubleCounter.bind(Attributes.builder().put("K", "V").build()))));
     }
 
     stressTestBuilder.build().run();
@@ -204,10 +209,10 @@ class DoubleUpDownCounterSdkTest {
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testUpDownCounter")
+                    .hasName("testCounter")
                     .hasDoubleSum()
                     .isCumulative()
-                    .isNotMonotonic()
+                    .isMonotonic()
                     .points()
                     .satisfiesExactlyInAnyOrder(
                         point ->
@@ -221,29 +226,27 @@ class DoubleUpDownCounterSdkTest {
   }
 
   @Test
-  @SuppressWarnings("unchecked")
   void stressTest_WithDifferentLabelSet() {
     final String[] keys = {"Key_1", "Key_2", "Key_3", "Key_4"};
     final String[] values = {"Value_1", "Value_2", "Value_3", "Value_4"};
-    final DoubleUpDownCounter doubleUpDownCounter =
-        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
+    final DoubleCounter doubleCounter = sdkMeter.counterBuilder("testCounter").ofDoubles().build();
 
     StressTestRunner.Builder stressTestBuilder =
         StressTestRunner.builder()
-            .setInstrument((DoubleUpDownCounterSdk) doubleUpDownCounter)
+            .setInstrument((SdkDoubleCounter) doubleCounter)
             .setCollectionIntervalMs(100);
 
     for (int i = 0; i < 4; i++) {
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
-              2_000, 1, new OperationUpdaterDirectCall(doubleUpDownCounter, keys[i], values[i])));
+              2_000, 1, new OperationUpdaterDirectCall(doubleCounter, keys[i], values[i])));
 
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
               2_000,
               1,
               new OperationUpdaterWithBinding(
-                  doubleUpDownCounter.bind(Attributes.builder().put(keys[i], values[i]).build()))));
+                  doubleCounter.bind(Attributes.builder().put(keys[i], values[i]).build()))));
     }
 
     stressTestBuilder.build().run();
@@ -253,10 +256,9 @@ class DoubleUpDownCounterSdkTest {
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testUpDownCounter")
                     .hasDoubleSum()
                     .isCumulative()
-                    .isNotMonotonic()
+                    .isMonotonic()
                     .points()
                     .allSatisfy(
                         point ->
@@ -273,39 +275,38 @@ class DoubleUpDownCounterSdkTest {
   }
 
   private static class OperationUpdaterWithBinding extends OperationUpdater {
-    private final BoundDoubleUpDownCounter boundDoubleUpDownCounter;
+    private final BoundDoubleCounter boundDoubleCounter;
 
-    private OperationUpdaterWithBinding(BoundDoubleUpDownCounter boundDoubleUpDownCounter) {
-      this.boundDoubleUpDownCounter = boundDoubleUpDownCounter;
+    private OperationUpdaterWithBinding(BoundDoubleCounter boundDoubleCounter) {
+      this.boundDoubleCounter = boundDoubleCounter;
     }
 
     @Override
     void update() {
-      boundDoubleUpDownCounter.add(9.0);
+      boundDoubleCounter.add(9.0);
     }
 
     @Override
     void cleanup() {
-      boundDoubleUpDownCounter.unbind();
+      boundDoubleCounter.unbind();
     }
   }
 
   private static class OperationUpdaterDirectCall extends OperationUpdater {
 
-    private final DoubleUpDownCounter doubleUpDownCounter;
+    private final DoubleCounter doubleCounter;
     private final String key;
     private final String value;
 
-    private OperationUpdaterDirectCall(
-        DoubleUpDownCounter doubleUpDownCounter, String key, String value) {
-      this.doubleUpDownCounter = doubleUpDownCounter;
+    private OperationUpdaterDirectCall(DoubleCounter doubleCounter, String key, String value) {
+      this.doubleCounter = doubleCounter;
       this.key = key;
       this.value = value;
     }
 
     @Override
     void update() {
-      doubleUpDownCounter.add(11.0, Attributes.builder().put(key, value).build());
+      doubleCounter.add(11.0, Attributes.builder().put(key, value).build());
     }
 
     @Override
