+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/SdkDoubleGaugeBuilderTest.java
@@ -16,12 +16,12 @@ import io.opentelemetry.sdk.testing.time.TestClock;
 import java.time.Duration;
 import org.junit.jupiter.api.Test;
 
-/** Unit tests for {@link LongValueObserverSdk}. */
-class LongGaugeBuilderSdkTest {
+/** Unit tests for {@link DoubleValueObserverSdk}. */
+class SdkDoubleGaugeBuilderTest {
   private static final Resource RESOURCE =
       Resource.create(Attributes.of(stringKey("resource_key"), "resource_value"));
   private static final InstrumentationLibraryInfo INSTRUMENTATION_LIBRARY_INFO =
-      InstrumentationLibraryInfo.create(LongGaugeBuilderSdkTest.class.getName(), null);
+      InstrumentationLibraryInfo.create(SdkDoubleGaugeBuilderTest.class.getName(), null);
   private final TestClock testClock = TestClock.create();
   private final SdkMeterProvider sdkMeterProvider =
       SdkMeterProvider.builder().setClock(testClock).setResource(RESOURCE).build();
@@ -31,8 +31,7 @@ class LongGaugeBuilderSdkTest {
   void collectMetrics_NoRecords() {
     sdkMeter
         .gaugeBuilder("testObserver")
-        .ofLongs()
-        .setDescription("My own LongValueObserver")
+        .setDescription("My own DoubleValueObserver")
         .setUnit("ms")
         .buildWithCallback(result -> {});
     assertThat(sdkMeterProvider.collectAllMetrics()).isEmpty();
@@ -43,9 +42,10 @@ class LongGaugeBuilderSdkTest {
   void collectMetrics_WithOneRecord() {
     sdkMeter
         .gaugeBuilder("testObserver")
-        .ofLongs()
+        .setDescription("My own DoubleValueObserver")
+        .setUnit("ms")
         .buildWithCallback(
-            result -> result.observe(12, Attributes.builder().put("k", "v").build()));
+            result -> result.observe(12.1d, Attributes.builder().put("k", "v").build()));
     testClock.advance(Duration.ofSeconds(1));
     assertThat(sdkMeterProvider.collectAllMetrics())
         .satisfiesExactly(
@@ -54,7 +54,9 @@ class LongGaugeBuilderSdkTest {
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
                     .hasName("testObserver")
-                    .hasLongGauge()
+                    .hasDescription("My own DoubleValueObserver")
+                    .hasUnit("ms")
+                    .hasDoubleGauge()
                     .points()
                     .satisfiesExactlyInAnyOrder(
                         point ->
@@ -62,7 +64,7 @@ class LongGaugeBuilderSdkTest {
                                 .hasStartEpochNanos(0)
                                 .hasEpochNanos(testClock.now())
                                 .hasAttributes(Attributes.builder().put("k", "v").build())
-                                .hasValue(12)));
+                                .hasValue(12.1d)));
     testClock.advance(Duration.ofSeconds(1));
     assertThat(sdkMeterProvider.collectAllMetrics())
         .satisfiesExactly(
@@ -71,7 +73,9 @@ class LongGaugeBuilderSdkTest {
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
                     .hasName("testObserver")
-                    .hasLongGauge()
+                    .hasDescription("My own DoubleValueObserver")
+                    .hasUnit("ms")
+                    .hasDoubleGauge()
                     .points()
                     .satisfiesExactlyInAnyOrder(
                         point ->
@@ -79,6 +83,6 @@ class LongGaugeBuilderSdkTest {
                                 .hasStartEpochNanos(0)
                                 .hasEpochNanos(testClock.now())
                                 .hasAttributes(Attributes.builder().put("k", "v").build())
-                                .hasValue(12)));
+                                .hasValue(12.1d)));
   }
 }
