+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/SdkDoubleHistogramTest.java
@@ -7,11 +7,12 @@ package io.opentelemetry.sdk.metrics;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 import static io.opentelemetry.sdk.testing.assertj.metrics.MetricAssertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.metrics.BoundLongCounter;
-import io.opentelemetry.api.metrics.LongCounter;
+import io.opentelemetry.api.metrics.BoundDoubleHistogram;
+import io.opentelemetry.api.metrics.DoubleHistogram;
 import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.metrics.StressTestRunner.OperationUpdater;
@@ -20,36 +21,37 @@ import io.opentelemetry.sdk.testing.time.TestClock;
 import java.time.Duration;
 import org.junit.jupiter.api.Test;
 
-/** Unit tests for {@link LongCounterSdk}. */
-class LongCounterSdkTest {
+/** Unit tests for {@link SdkDoubleHistogram}. */
+class SdkDoubleHistogramTest {
   private static final long SECOND_NANOS = 1_000_000_000;
   private static final Resource RESOURCE =
       Resource.create(Attributes.of(stringKey("resource_key"), "resource_value"));
   private static final InstrumentationLibraryInfo INSTRUMENTATION_LIBRARY_INFO =
-      InstrumentationLibraryInfo.create(LongCounterSdkTest.class.getName(), null);
+      InstrumentationLibraryInfo.create(SdkDoubleHistogramTest.class.getName(), null);
   private final TestClock testClock = TestClock.create();
   private final SdkMeterProvider sdkMeterProvider =
       SdkMeterProvider.builder().setClock(testClock).setResource(RESOURCE).build();
   private final Meter sdkMeter = sdkMeterProvider.get(getClass().getName());
 
   @Test
-  void add_PreventNullAttributes() {
-    assertThatThrownBy(() -> sdkMeter.counterBuilder("testCounter").build().add(1, null))
+  void record_PreventNullAttributes() {
+    assertThatThrownBy(() -> sdkMeter.histogramBuilder("testRecorder").build().record(1.0, null))
         .isInstanceOf(NullPointerException.class)
         .hasMessage("attributes");
   }
 
   @Test
   void bound_PreventNullAttributes() {
-    assertThatThrownBy(() -> sdkMeter.counterBuilder("testCounter").build().bind(null))
+    assertThatThrownBy(() -> sdkMeter.histogramBuilder("testRecorder").build().bind(null))
         .isInstanceOf(NullPointerException.class)
         .hasMessage("attributes");
   }
 
   @Test
   void collectMetrics_NoRecords() {
-    LongCounter longCounter = sdkMeter.counterBuilder("Counter").build();
-    BoundLongCounter bound = longCounter.bind(Attributes.builder().put("foo", "bar").build());
+    DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
+    BoundDoubleHistogram bound =
+        doubleRecorder.bind(Attributes.builder().put("key", "value").build());
     try {
       assertThat(sdkMeterProvider.collectAllMetrics()).isEmpty();
     } finally {
@@ -58,24 +60,26 @@ class LongCounterSdkTest {
   }
 
   @Test
-  @SuppressWarnings("unchecked")
   void collectMetrics_WithEmptyAttributes() {
-    LongCounter longCounter =
-        sdkMeter.counterBuilder("testCounter").setDescription("description").setUnit("By").build();
+    DoubleHistogram doubleRecorder =
+        sdkMeter
+            .histogramBuilder("testRecorder")
+            .setDescription("description")
+            .setUnit("ms")
+            .build();
     testClock.advance(Duration.ofNanos(SECOND_NANOS));
-    longCounter.add(12, Attributes.empty());
-    longCounter.add(12);
+    doubleRecorder.record(12d, Attributes.empty());
+    doubleRecorder.record(12d);
     assertThat(sdkMeterProvider.collectAllMetrics())
         .satisfiesExactly(
             metric ->
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testCounter")
+                    .hasName("testRecorder")
                     .hasDescription("description")
-                    .hasUnit("By")
-                    .hasLongSum()
-                    .isMonotonic()
+                    .hasUnit("ms")
+                    .hasDoubleHistogram()
                     .isCumulative()
                     .points()
                     .satisfiesExactly(
@@ -84,34 +88,37 @@ class LongCounterSdkTest {
                                 .hasStartEpochNanos(testClock.now() - SECOND_NANOS)
                                 .hasEpochNanos(testClock.now())
                                 .hasAttributes(Attributes.empty())
-                                .hasValue(24)));
+                                .hasCount(2)
+                                .hasSum(24)
+                                .hasBucketBoundaries(
+                                    5, 10, 25, 50, 75, 100, 250, 500, 750, 1_000, 2_500, 5_000,
+                                    7_500, 10_000)
+                                .hasBucketCounts(0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)));
   }
 
   @Test
   @SuppressWarnings("unchecked")
   void collectMetrics_WithMultipleCollects() {
     long startTime = testClock.now();
-    LongCounter longCounter = sdkMeter.counterBuilder("testCounter").build();
-    BoundLongCounter bound = longCounter.bind(Attributes.builder().put("K", "V").build());
+    DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
+    BoundDoubleHistogram bound = doubleRecorder.bind(Attributes.builder().put("K", "V").build());
     try {
       // Do some records using bounds and direct calls and bindings.
-      longCounter.add(12, Attributes.empty());
-      bound.add(123);
-      longCounter.add(21, Attributes.empty());
+      doubleRecorder.record(12.1d, Attributes.empty());
+      bound.record(123.3d);
+      doubleRecorder.record(-13.1d, Attributes.empty());
       // Advancing time here should not matter.
       testClock.advance(Duration.ofNanos(SECOND_NANOS));
-      bound.add(321);
-      longCounter.add(111, Attributes.builder().put("K", "V").build());
+      bound.record(321.5d);
+      doubleRecorder.record(-121.5d, Attributes.builder().put("K", "V").build());
       assertThat(sdkMeterProvider.collectAllMetrics())
           .satisfiesExactly(
               metric ->
                   assertThat(metric)
                       .hasResource(RESOURCE)
                       .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                      .hasName("testCounter")
-                      .hasLongSum()
-                      .isMonotonic()
-                      .isCumulative()
+                      .hasName("testRecorder")
+                      .hasDoubleHistogram()
                       .points()
                       .allSatisfy(
                           point ->
@@ -119,26 +126,31 @@ class LongCounterSdkTest {
                                   .hasStartEpochNanos(startTime)
                                   .hasEpochNanos(testClock.now()))
                       .satisfiesExactlyInAnyOrder(
-                          point -> assertThat(point).hasAttributes(Attributes.empty()).hasValue(33),
                           point ->
                               assertThat(point)
-                                  .hasAttributes(Attributes.of(stringKey("K"), "V"))
-                                  .hasValue(555)));
+                                  .hasCount(3)
+                                  .hasSum(323.3d)
+                                  .hasBucketCounts(1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0)
+                                  .hasAttributes(Attributes.builder().put("K", "V").build()),
+                          point ->
+                              assertThat(point)
+                                  .hasCount(2)
+                                  .hasSum(-1.0d)
+                                  .hasBucketCounts(1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
+                                  .hasAttributes(Attributes.empty())));
 
-      // Repeat to prove we keep previous values.
+      // Histograms are cumulative by default.
       testClock.advance(Duration.ofNanos(SECOND_NANOS));
-      bound.add(222);
-      longCounter.add(11, Attributes.empty());
+      bound.record(222d);
+      doubleRecorder.record(17d, Attributes.empty());
       assertThat(sdkMeterProvider.collectAllMetrics())
           .satisfiesExactly(
               metric ->
                   assertThat(metric)
                       .hasResource(RESOURCE)
                       .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                      .hasName("testCounter")
-                      .hasLongSum()
-                      .isMonotonic()
-                      .isCumulative()
+                      .hasName("testRecorder")
+                      .hasDoubleHistogram()
                       .points()
                       .allSatisfy(
                           point ->
@@ -146,100 +158,44 @@ class LongCounterSdkTest {
                                   .hasStartEpochNanos(startTime)
                                   .hasEpochNanos(testClock.now()))
                       .satisfiesExactlyInAnyOrder(
-                          point -> assertThat(point).hasAttributes(Attributes.empty()).hasValue(44),
                           point ->
                               assertThat(point)
-                                  .hasAttributes(Attributes.of(stringKey("K"), "V"))
-                                  .hasValue(777)));
+                                  .hasCount(4)
+                                  .hasSum(545.3)
+                                  .hasBucketCounts(1, 0, 0, 0, 0, 0, 2, 1, 0, 0, 0, 0, 0, 0, 0)
+                                  .hasAttributes(Attributes.builder().put("K", "V").build()),
+                          point ->
+                              assertThat(point)
+                                  .hasCount(3)
+                                  .hasSum(16)
+                                  .hasBucketCounts(1, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
+                                  .hasAttributes(Attributes.empty())));
     } finally {
       bound.unbind();
     }
   }
 
   @Test
-  void longCounterAdd_MonotonicityCheck() {
-    LongCounter longCounter = sdkMeter.counterBuilder("testCounter").build();
-
-    assertThatThrownBy(() -> longCounter.add(-45, Attributes.empty()))
-        .isInstanceOf(IllegalArgumentException.class);
-  }
-
-  @Test
-  void boundLongCounterAdd_MonotonicityCheck() {
-    LongCounter longCounter = sdkMeter.counterBuilder("testCounter").build();
-
-    assertThatThrownBy(() -> longCounter.bind(Attributes.empty()).add(-9))
-        .isInstanceOf(IllegalArgumentException.class);
-  }
-
-  @Test
-  @SuppressWarnings("unchecked")
   void stressTest() {
-    final LongCounter longCounter = sdkMeter.counterBuilder("testCounter").build();
+    final DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
 
     StressTestRunner.Builder stressTestBuilder =
         StressTestRunner.builder()
-            .setInstrument((LongCounterSdk) longCounter)
+            .setInstrument((SdkDoubleHistogram) doubleRecorder)
             .setCollectionIntervalMs(100);
 
     for (int i = 0; i < 4; i++) {
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
-              2_000, 1, new OperationUpdaterDirectCall(longCounter, "K", "V")));
-      stressTestBuilder.addOperation(
-          StressTestRunner.Operation.create(
-              2_000,
-              1,
-              new OperationUpdaterWithBinding(
-                  longCounter.bind(Attributes.builder().put("K", "V").build()))));
-    }
-
-    stressTestBuilder.build().run();
-    assertThat(sdkMeterProvider.collectAllMetrics())
-        .satisfiesExactly(
-            metric ->
-                assertThat(metric)
-                    .hasResource(RESOURCE)
-                    .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testCounter")
-                    .hasLongSum()
-                    .isCumulative()
-                    .isMonotonic()
-                    .points()
-                    .satisfiesExactlyInAnyOrder(
-                        point ->
-                            assertThat(point)
-                                .hasStartEpochNanos(testClock.now())
-                                .hasEpochNanos(testClock.now())
-                                .hasValue(160_000)
-                                .attributes()
-                                .hasSize(1)
-                                .containsEntry("K", "V")));
-  }
-
-  @Test
-  @SuppressWarnings("unchecked")
-  void stressTest_WithDifferentLabelSet() {
-    final String[] keys = {"Key_1", "Key_2", "Key_3", "Key_4"};
-    final String[] values = {"Value_1", "Value_2", "Value_3", "Value_4"};
-    final LongCounter longCounter = sdkMeter.counterBuilder("testCounter").build();
-
-    StressTestRunner.Builder stressTestBuilder =
-        StressTestRunner.builder()
-            .setInstrument((LongCounterSdk) longCounter)
-            .setCollectionIntervalMs(100);
-
-    for (int i = 0; i < 4; i++) {
-      stressTestBuilder.addOperation(
-          StressTestRunner.Operation.create(
-              1_000, 2, new OperationUpdaterDirectCall(longCounter, keys[i], values[i])));
-
+              1_000,
+              2,
+              new SdkDoubleHistogramTest.OperationUpdaterDirectCall(doubleRecorder, "K", "V")));
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
               1_000,
               2,
               new OperationUpdaterWithBinding(
-                  longCounter.bind(Attributes.builder().put(keys[i], values[i]).build()))));
+                  doubleRecorder.bind(Attributes.builder().put("K", "V").build()))));
     }
 
     stressTestBuilder.build().run();
@@ -249,17 +205,64 @@ class LongCounterSdkTest {
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testCounter")
-                    .hasLongSum()
-                    .isCumulative()
-                    .isMonotonic()
+                    .hasName("testRecorder")
+                    .hasDoubleHistogram()
+                    .points()
+                    .satisfiesExactly(
+                        point ->
+                            assertThat(point)
+                                .hasStartEpochNanos(testClock.now())
+                                .hasEpochNanos(testClock.now())
+                                .hasAttributes(Attributes.of(stringKey("K"), "V"))
+                                .hasCount(8_000)
+                                .hasSum(80_000)));
+  }
+
+  @Test
+  void stressTest_WithDifferentLabelSet() {
+    final String[] keys = {"Key_1", "Key_2", "Key_3", "Key_4"};
+    final String[] values = {"Value_1", "Value_2", "Value_3", "Value_4"};
+    final DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
+
+    StressTestRunner.Builder stressTestBuilder =
+        StressTestRunner.builder()
+            .setInstrument((SdkDoubleHistogram) doubleRecorder)
+            .setCollectionIntervalMs(100);
+
+    for (int i = 0; i < 4; i++) {
+      stressTestBuilder.addOperation(
+          StressTestRunner.Operation.create(
+              2_000,
+              1,
+              new SdkDoubleHistogramTest.OperationUpdaterDirectCall(
+                  doubleRecorder, keys[i], values[i])));
+
+      stressTestBuilder.addOperation(
+          StressTestRunner.Operation.create(
+              2_000,
+              1,
+              new OperationUpdaterWithBinding(
+                  doubleRecorder.bind(Attributes.builder().put(keys[i], values[i]).build()))));
+    }
+
+    stressTestBuilder.build().run();
+    assertThat(sdkMeterProvider.collectAllMetrics())
+        .satisfiesExactly(
+            metric ->
+                assertThat(metric)
+                    .hasResource(RESOURCE)
+                    .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
+                    .hasName("testRecorder")
+                    .hasDoubleHistogram()
                     .points()
                     .allSatisfy(
                         point ->
                             assertThat(point)
                                 .hasStartEpochNanos(testClock.now())
                                 .hasEpochNanos(testClock.now())
-                                .hasValue(20_000))
+                                .hasCount(4_000)
+                                .hasSum(40_000)
+                                .hasBucketCounts(0, 2000, 2000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0))
                     .extracting(point -> point.getAttributes())
                     .containsExactlyInAnyOrder(
                         Attributes.of(stringKey(keys[0]), values[0]),
@@ -269,38 +272,38 @@ class LongCounterSdkTest {
   }
 
   private static class OperationUpdaterWithBinding extends OperationUpdater {
-    private final BoundLongCounter boundLongCounter;
+    private final BoundDoubleHistogram boundDoubleValueRecorder;
 
-    private OperationUpdaterWithBinding(BoundLongCounter boundLongCounter) {
-      this.boundLongCounter = boundLongCounter;
+    private OperationUpdaterWithBinding(BoundDoubleHistogram boundDoubleValueRecorder) {
+      this.boundDoubleValueRecorder = boundDoubleValueRecorder;
     }
 
     @Override
     void update() {
-      boundLongCounter.add(9);
+      boundDoubleValueRecorder.record(11.0);
     }
 
     @Override
     void cleanup() {
-      boundLongCounter.unbind();
+      boundDoubleValueRecorder.unbind();
     }
   }
 
   private static class OperationUpdaterDirectCall extends OperationUpdater {
-
-    private final LongCounter longCounter;
+    private final DoubleHistogram doubleValueRecorder;
     private final String key;
     private final String value;
 
-    private OperationUpdaterDirectCall(LongCounter longCounter, String key, String value) {
-      this.longCounter = longCounter;
+    private OperationUpdaterDirectCall(
+        DoubleHistogram doubleValueRecorder, String key, String value) {
+      this.doubleValueRecorder = doubleValueRecorder;
       this.key = key;
       this.value = value;
     }
 
     @Override
     void update() {
-      longCounter.add(11, Attributes.builder().put(key, value).build());
+      doubleValueRecorder.record(9.0, Attributes.builder().put(key, value).build());
     }
 
     @Override
