+++ sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/SdkDoubleUpDownCounterTest.java
@@ -7,12 +7,11 @@ package io.opentelemetry.sdk.metrics;
 
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 import static io.opentelemetry.sdk.testing.assertj.metrics.MetricAssertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.metrics.BoundDoubleHistogram;
-import io.opentelemetry.api.metrics.DoubleHistogram;
+import io.opentelemetry.api.metrics.BoundDoubleUpDownCounter;
+import io.opentelemetry.api.metrics.DoubleUpDownCounter;
 import io.opentelemetry.api.metrics.Meter;
 import io.opentelemetry.sdk.common.InstrumentationLibraryInfo;
 import io.opentelemetry.sdk.metrics.StressTestRunner.OperationUpdater;
@@ -21,37 +20,45 @@ import io.opentelemetry.sdk.testing.time.TestClock;
 import java.time.Duration;
 import org.junit.jupiter.api.Test;
 
-/** Unit tests for {@link DoubleHistogramSdk}. */
-class DoubleHistogramSdkTest {
+/** Unit tests for {@link SdkDoubleUpDownCounter}. */
+class SdkDoubleUpDownCounterTest {
   private static final long SECOND_NANOS = 1_000_000_000;
   private static final Resource RESOURCE =
       Resource.create(Attributes.of(stringKey("resource_key"), "resource_value"));
   private static final InstrumentationLibraryInfo INSTRUMENTATION_LIBRARY_INFO =
-      InstrumentationLibraryInfo.create(DoubleHistogramSdkTest.class.getName(), null);
+      InstrumentationLibraryInfo.create(SdkDoubleUpDownCounterTest.class.getName(), null);
   private final TestClock testClock = TestClock.create();
   private final SdkMeterProvider sdkMeterProvider =
       SdkMeterProvider.builder().setClock(testClock).setResource(RESOURCE).build();
   private final Meter sdkMeter = sdkMeterProvider.get(getClass().getName());
 
   @Test
-  void record_PreventNullAttributes() {
-    assertThatThrownBy(() -> sdkMeter.histogramBuilder("testRecorder").build().record(1.0, null))
+  void add_PreventNullAttributes() {
+    assertThatThrownBy(
+            () ->
+                sdkMeter
+                    .upDownCounterBuilder("testUpDownCounter")
+                    .ofDoubles()
+                    .build()
+                    .add(1.0, null))
         .isInstanceOf(NullPointerException.class)
         .hasMessage("attributes");
   }
 
   @Test
   void bound_PreventNullAttributes() {
-    assertThatThrownBy(() -> sdkMeter.histogramBuilder("testRecorder").build().bind(null))
+    assertThatThrownBy(
+            () -> sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build().bind(null))
         .isInstanceOf(NullPointerException.class)
         .hasMessage("attributes");
   }
 
   @Test
   void collectMetrics_NoRecords() {
-    DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
-    BoundDoubleHistogram bound =
-        doubleRecorder.bind(Attributes.builder().put("key", "value").build());
+    DoubleUpDownCounter doubleUpDownCounter =
+        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
+    BoundDoubleUpDownCounter bound =
+        doubleUpDownCounter.bind(Attributes.builder().put("foo", "bar").build());
     try {
       assertThat(sdkMeterProvider.collectAllMetrics()).isEmpty();
     } finally {
@@ -60,26 +67,29 @@ class DoubleHistogramSdkTest {
   }
 
   @Test
+  @SuppressWarnings("unchecked")
   void collectMetrics_WithEmptyAttributes() {
-    DoubleHistogram doubleRecorder =
+    DoubleUpDownCounter doubleUpDownCounter =
         sdkMeter
-            .histogramBuilder("testRecorder")
+            .upDownCounterBuilder("testUpDownCounter")
+            .ofDoubles()
             .setDescription("description")
             .setUnit("ms")
             .build();
     testClock.advance(Duration.ofNanos(SECOND_NANOS));
-    doubleRecorder.record(12d, Attributes.empty());
-    doubleRecorder.record(12d);
+    doubleUpDownCounter.add(12d, Attributes.empty());
+    doubleUpDownCounter.add(12d);
     assertThat(sdkMeterProvider.collectAllMetrics())
         .satisfiesExactly(
             metric ->
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testRecorder")
+                    .hasName("testUpDownCounter")
                     .hasDescription("description")
                     .hasUnit("ms")
-                    .hasDoubleHistogram()
+                    .hasDoubleSum()
+                    .isNotMonotonic()
                     .isCumulative()
                     .points()
                     .satisfiesExactly(
@@ -88,37 +98,36 @@ class DoubleHistogramSdkTest {
                                 .hasStartEpochNanos(testClock.now() - SECOND_NANOS)
                                 .hasEpochNanos(testClock.now())
                                 .hasAttributes(Attributes.empty())
-                                .hasCount(2)
-                                .hasSum(24)
-                                .hasBucketBoundaries(
-                                    5, 10, 25, 50, 75, 100, 250, 500, 750, 1_000, 2_500, 5_000,
-                                    7_500, 10_000)
-                                .hasBucketCounts(0, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)));
+                                .hasValue(24)));
   }
 
   @Test
   @SuppressWarnings("unchecked")
   void collectMetrics_WithMultipleCollects() {
     long startTime = testClock.now();
-    DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
-    BoundDoubleHistogram bound = doubleRecorder.bind(Attributes.builder().put("K", "V").build());
+    DoubleUpDownCounter doubleUpDownCounter =
+        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
+    BoundDoubleUpDownCounter bound =
+        doubleUpDownCounter.bind(Attributes.builder().put("K", "V").build());
     try {
       // Do some records using bounds and direct calls and bindings.
-      doubleRecorder.record(12.1d, Attributes.empty());
-      bound.record(123.3d);
-      doubleRecorder.record(-13.1d, Attributes.empty());
+      doubleUpDownCounter.add(12.1d, Attributes.empty());
+      bound.add(123.3d);
+      doubleUpDownCounter.add(21.4d, Attributes.empty());
       // Advancing time here should not matter.
       testClock.advance(Duration.ofNanos(SECOND_NANOS));
-      bound.record(321.5d);
-      doubleRecorder.record(-121.5d, Attributes.builder().put("K", "V").build());
+      bound.add(321.5d);
+      doubleUpDownCounter.add(111.1d, Attributes.builder().put("K", "V").build());
       assertThat(sdkMeterProvider.collectAllMetrics())
           .satisfiesExactly(
               metric ->
                   assertThat(metric)
                       .hasResource(RESOURCE)
                       .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                      .hasName("testRecorder")
-                      .hasDoubleHistogram()
+                      .hasName("testUpDownCounter")
+                      .hasDoubleSum()
+                      .isNotMonotonic()
+                      .isCumulative()
                       .points()
                       .allSatisfy(
                           point ->
@@ -127,30 +136,26 @@ class DoubleHistogramSdkTest {
                                   .hasEpochNanos(testClock.now()))
                       .satisfiesExactlyInAnyOrder(
                           point ->
-                              assertThat(point)
-                                  .hasCount(3)
-                                  .hasSum(323.3d)
-                                  .hasBucketCounts(1, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0)
-                                  .hasAttributes(Attributes.builder().put("K", "V").build()),
+                              assertThat(point).hasAttributes(Attributes.empty()).hasValue(33.5),
                           point ->
                               assertThat(point)
-                                  .hasCount(2)
-                                  .hasSum(-1.0d)
-                                  .hasBucketCounts(1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
-                                  .hasAttributes(Attributes.empty())));
+                                  .hasValue(555.9)
+                                  .hasAttributes(Attributes.of(stringKey("K"), "V"))));
 
-      // Histograms are cumulative by default.
+      // Repeat to prove we keep previous values.
       testClock.advance(Duration.ofNanos(SECOND_NANOS));
-      bound.record(222d);
-      doubleRecorder.record(17d, Attributes.empty());
+      bound.add(222d);
+      doubleUpDownCounter.add(11d, Attributes.empty());
       assertThat(sdkMeterProvider.collectAllMetrics())
           .satisfiesExactly(
               metric ->
                   assertThat(metric)
                       .hasResource(RESOURCE)
                       .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                      .hasName("testRecorder")
-                      .hasDoubleHistogram()
+                      .hasName("testUpDownCounter")
+                      .hasDoubleSum()
+                      .isNotMonotonic()
+                      .isCumulative()
                       .points()
                       .allSatisfy(
                           point ->
@@ -159,43 +164,37 @@ class DoubleHistogramSdkTest {
                                   .hasEpochNanos(testClock.now()))
                       .satisfiesExactlyInAnyOrder(
                           point ->
-                              assertThat(point)
-                                  .hasCount(4)
-                                  .hasSum(545.3)
-                                  .hasBucketCounts(1, 0, 0, 0, 0, 0, 2, 1, 0, 0, 0, 0, 0, 0, 0)
-                                  .hasAttributes(Attributes.builder().put("K", "V").build()),
+                              assertThat(point).hasAttributes(Attributes.empty()).hasValue(44.5),
                           point ->
                               assertThat(point)
-                                  .hasCount(3)
-                                  .hasSum(16)
-                                  .hasBucketCounts(1, 0, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
-                                  .hasAttributes(Attributes.empty())));
+                                  .hasAttributes(Attributes.of(stringKey("K"), "V"))
+                                  .hasValue(777.9)));
     } finally {
       bound.unbind();
     }
   }
 
   @Test
+  @SuppressWarnings("unchecked")
   void stressTest() {
-    final DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
+    final DoubleUpDownCounter doubleUpDownCounter =
+        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
 
     StressTestRunner.Builder stressTestBuilder =
         StressTestRunner.builder()
-            .setInstrument((DoubleHistogramSdk) doubleRecorder)
+            .setInstrument((SdkDoubleUpDownCounter) doubleUpDownCounter)
             .setCollectionIntervalMs(100);
 
     for (int i = 0; i < 4; i++) {
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
-              1_000,
-              2,
-              new DoubleHistogramSdkTest.OperationUpdaterDirectCall(doubleRecorder, "K", "V")));
+              1_000, 2, new OperationUpdaterDirectCall(doubleUpDownCounter, "K", "V")));
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
               1_000,
               2,
               new OperationUpdaterWithBinding(
-                  doubleRecorder.bind(Attributes.builder().put("K", "V").build()))));
+                  doubleUpDownCounter.bind(Attributes.builder().put("K", "V").build()))));
     }
 
     stressTestBuilder.build().run();
@@ -205,44 +204,46 @@ class DoubleHistogramSdkTest {
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testRecorder")
-                    .hasDoubleHistogram()
+                    .hasName("testUpDownCounter")
+                    .hasDoubleSum()
+                    .isCumulative()
+                    .isNotMonotonic()
                     .points()
-                    .satisfiesExactly(
+                    .satisfiesExactlyInAnyOrder(
                         point ->
                             assertThat(point)
                                 .hasStartEpochNanos(testClock.now())
                                 .hasEpochNanos(testClock.now())
-                                .hasAttributes(Attributes.of(stringKey("K"), "V"))
-                                .hasCount(8_000)
-                                .hasSum(80_000)));
+                                .hasValue(80_000)
+                                .attributes()
+                                .hasSize(1)
+                                .containsEntry("K", "V")));
   }
 
   @Test
+  @SuppressWarnings("unchecked")
   void stressTest_WithDifferentLabelSet() {
     final String[] keys = {"Key_1", "Key_2", "Key_3", "Key_4"};
     final String[] values = {"Value_1", "Value_2", "Value_3", "Value_4"};
-    final DoubleHistogram doubleRecorder = sdkMeter.histogramBuilder("testRecorder").build();
+    final DoubleUpDownCounter doubleUpDownCounter =
+        sdkMeter.upDownCounterBuilder("testUpDownCounter").ofDoubles().build();
 
     StressTestRunner.Builder stressTestBuilder =
         StressTestRunner.builder()
-            .setInstrument((DoubleHistogramSdk) doubleRecorder)
+            .setInstrument((SdkDoubleUpDownCounter) doubleUpDownCounter)
             .setCollectionIntervalMs(100);
 
     for (int i = 0; i < 4; i++) {
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
-              2_000,
-              1,
-              new DoubleHistogramSdkTest.OperationUpdaterDirectCall(
-                  doubleRecorder, keys[i], values[i])));
+              2_000, 1, new OperationUpdaterDirectCall(doubleUpDownCounter, keys[i], values[i])));
 
       stressTestBuilder.addOperation(
           StressTestRunner.Operation.create(
               2_000,
               1,
               new OperationUpdaterWithBinding(
-                  doubleRecorder.bind(Attributes.builder().put(keys[i], values[i]).build()))));
+                  doubleUpDownCounter.bind(Attributes.builder().put(keys[i], values[i]).build()))));
     }
 
     stressTestBuilder.build().run();
@@ -252,17 +253,17 @@ class DoubleHistogramSdkTest {
                 assertThat(metric)
                     .hasResource(RESOURCE)
                     .hasInstrumentationLibrary(INSTRUMENTATION_LIBRARY_INFO)
-                    .hasName("testRecorder")
-                    .hasDoubleHistogram()
+                    .hasName("testUpDownCounter")
+                    .hasDoubleSum()
+                    .isCumulative()
+                    .isNotMonotonic()
                     .points()
                     .allSatisfy(
                         point ->
                             assertThat(point)
                                 .hasStartEpochNanos(testClock.now())
                                 .hasEpochNanos(testClock.now())
-                                .hasCount(4_000)
-                                .hasSum(40_000)
-                                .hasBucketCounts(0, 2000, 2000, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0))
+                                .hasValue(40_000))
                     .extracting(point -> point.getAttributes())
                     .containsExactlyInAnyOrder(
                         Attributes.of(stringKey(keys[0]), values[0]),
@@ -272,38 +273,39 @@ class DoubleHistogramSdkTest {
   }
 
   private static class OperationUpdaterWithBinding extends OperationUpdater {
-    private final BoundDoubleHistogram boundDoubleValueRecorder;
+    private final BoundDoubleUpDownCounter boundDoubleUpDownCounter;
 
-    private OperationUpdaterWithBinding(BoundDoubleHistogram boundDoubleValueRecorder) {
-      this.boundDoubleValueRecorder = boundDoubleValueRecorder;
+    private OperationUpdaterWithBinding(BoundDoubleUpDownCounter boundDoubleUpDownCounter) {
+      this.boundDoubleUpDownCounter = boundDoubleUpDownCounter;
     }
 
     @Override
     void update() {
-      boundDoubleValueRecorder.record(11.0);
+      boundDoubleUpDownCounter.add(9.0);
     }
 
     @Override
     void cleanup() {
-      boundDoubleValueRecorder.unbind();
+      boundDoubleUpDownCounter.unbind();
     }
   }
 
   private static class OperationUpdaterDirectCall extends OperationUpdater {
-    private final DoubleHistogram doubleValueRecorder;
+
+    private final DoubleUpDownCounter doubleUpDownCounter;
     private final String key;
     private final String value;
 
     private OperationUpdaterDirectCall(
-        DoubleHistogram doubleValueRecorder, String key, String value) {
-      this.doubleValueRecorder = doubleValueRecorder;
+        DoubleUpDownCounter doubleUpDownCounter, String key, String value) {
+      this.doubleUpDownCounter = doubleUpDownCounter;
       this.key = key;
       this.value = value;
     }
 
     @Override
     void update() {
-      doubleValueRecorder.record(9.0, Attributes.builder().put(key, value).build());
+      doubleUpDownCounter.add(11.0, Attributes.builder().put(key, value).build());
     }
 
     @Override
